+++ spec/src/qlib/UtilSpec.ts
